+++ packages/app/src/components/atoms/Input.ts
@@ -6,7 +6,7 @@ interface InputProps {
 
 const Input = styled.input<InputProps>`
 	background-color: ${p => p.theme.ui.surface};
-	color: ${p => p.theme.ui.textOnSurfaceBackground};
+	color: ${p => p.theme.ui.textMinor};
 	border: 1px solid ${p => p.theme.ui.primaryFill};
 	width: 100%;
 	box-sizing: border-box;
@@ -18,14 +18,18 @@ const Input = styled.input<InputProps>`
 
 export const Select = styled.select<InputProps>`
 	background-color: ${p => p.theme.ui.surface};
-	color: ${p => p.theme.ui.textOnSurfaceBackground};
-	border: 1px solid ${p => p.theme.ui.primaryFill};
+	color: ${p => p.theme.ui.textMinor};
+	border: 1px solid ${p => p.theme.ui.backgroundBorderSeparator};
 	width: 100%;
 	box-sizing: border-box;
 
 	font-size: ${p => (p.beakSize || 'md') === 'md' ? '15px' : '13px'};
 	padding: ${p => (p.beakSize || 'md') === 'md' ? '3px 5px' : '2px 3px'};
 	border-radius: ${p => (p.beakSize || 'md') === 'md' ? '4px' : '3px'};
+
+	&:active:not(:disabled) {
+		border: 1px solid ${p => p.theme.ui.primaryFill};
+	}
 `;
 
 export const InputInvalidText = styled.span`
