+++ packages/app/src/features/realtime-value-editor/components/atoms/Form.ts
@@ -1,0 +1,19 @@
+import styled from 'styled-components';
+
+export const FormGroup = styled.div`
+	margin-bottom: 8px;
+
+	> div > article {
+		font-size: 13px;
+		padding: 3px 5px;
+		padding-bottom: 4px;
+		border-radius: 3px;
+	}
+`;
+
+export const Label = styled.label`
+	display: block;
+	margin-bottom: 4px;
+
+	font-size: 13px;
+`;
